+++ Cargo.toml
@@ -31,6 +31,7 @@ solders-traits = { workspace = true }
 solders-macros = { workspace = true }
 solders-primitives = { workspace = true }
 solders-account-decoder = { workspace = true }
+solders-transaction-status = { workspace = true }
 solana-program-test = { workspace = true, optional = true }
 solana-banks-client = { workspace = true, optional = true }
 solana-banks-interface = { version = "1.15.2", optional = true}
@@ -49,13 +50,14 @@ bankrun = [
 
 
 [workspace]
-members = [".", "traits", "macros", "primitives", "account-decoder"]
+members = [".", "traits", "macros", "primitives", "account-decoder", "transaction-status"]
 
 [workspace.dependencies]
 pyo3 = { version = "0.18.0", default-features = false }
 solders-macros = { path = "./macros", version = "0.14.4" }
 solders-traits = { path = "./traits", version = "0.14.4" }
 solders-account-decoder = { path = "./account-decoder", version = "0.14.4" }
+solders-transaction-status = { path = "./transaction-status", version = "0.14.4" }
 serde_with = { version = "^1.14.0", features = ["base64"] }
 solders-primitives = { path = "./primitives", version = "0.14.4" }
 solana-sdk = "^1.15.2"
